Add a new field 'email' to Accounts.
The email field is used when sending emails. In the future, it should
be possible to select a default account (and email). This addresse
will be the one visible at the top of the page. For the moment just
show the first one.
Also add the posibility to edit name & email in account configuration.